Make interaction tests for UnderlineNav
