+++ src/UnderlineNav2/examples.stories.tsx
@@ -4,11 +4,9 @@ import {Meta} from '@storybook/react'
 import UnderlineNav, {UnderlineNavProps} from './index'
 import CounterLabel from '../CounterLabel'
 import {BaseStyles, ThemeProvider} from '..'
-import {within, userEvent} from '@storybook/testing-library'
-import {expect} from '@storybook/jest'
 
 export default {
-  title: 'Layout/UnderlineNav',
+  title: 'Layout/UnderlineNav/examples',
   argTypes: {
     align: {
       defaultValue: 'left',
@@ -39,19 +37,6 @@ export const DefaultNav = (args: UnderlineNavProps) => {
       <UnderlineNav.Link>Item 3</UnderlineNav.Link>
     </UnderlineNav>
   )
-}
-
-// Testing selection
-DefaultNav.play = async ({canvasElement}: {canvasElement: HTMLElement}) => {
-  const canvas = within(canvasElement)
-  const item = await canvas.getByText('Item 1')
-  const aElem = await item.closest('a')
-  const item2 = await canvas.getByText('Item 2')
-  const aElem2 = await item2.closest('a')
-  userEvent.click(aElem2 as Element)
-  expect(aElem2).toHaveAttribute('aria-current', 'page')
-  userEvent.click(aElem as Element)
-  expect(aElem).toHaveAttribute('aria-current', 'page')
 }
 
 export const withIcons = (args: UnderlineNavProps) => {
